items now have effects. Player and Monster now inherit common functionality from Character class.
this is so I could avoid having to rewrite effect-related (and other) methods for both Monster and Player. Now methods can take both Player and Method as a Character parameter and act accordingly based on type. Not sure if this is good design.
Currently effects keep stacking due to a mistake in the ordering if/else's in Character.UpdateEffects, and also conflicting usage of the EffectIsApplied field in this function as well as in the Effect class. My brain is too fried to solve that now. Also  I will take another look over the structure of the project later to make things less messy.