fixed exponential increase of count due to recursion
the design here is horrendous! especially now that we need to copy the Item in the BaseItems dict (to avoid recursion otherwise caused by the unintended updating of the Item in BaseItems dict because objects are passed by reference)
I'll have to find an alternative to the static dictionary if we want to make it prettier.